Remove my association from project
Though I authored the initial internal versions of this
(client-side) code, I've been told by many that I should be careful
of where I allow my name, because, as a security professional, the
link with it could possibly have some undefined potential to be not
nice for reputation. I thus take the advice of my peers and remove
my name from this codebase.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>