Add support to search using regexp
All commands that accept a UNIQUE_NAME parameter accept regular
expressions names as well, so instead of writing:

    lpass show 'Generated Password for soundcloud.com'

you can call

    lpass show /soundcl/

as well. When multiple accounts are matched, execution aborts with a
multiple match error (as if mulitple accounts had the same name)

This is an alternative to lastpass/lastpass-cli#37 and was developed in
parallel. I would welcome one of those PRs being supported.

Signed-off-by: Daniel Bornkessel <daniel@bornkessel.com>