http: add a method for dynamic parameter lists
Currently, the set of parameters must be known ahead of time,
or else http_post_lastpass_v must be used on an argv array
that is built up accordingly.

Add a new function, http_post_lastpass_params(), that can dynamically build
this argv array using a parameter list, and add a version of
http_post_lastpass() that calls http_post_lastpass_v on the end result.

Signed-off-by: Bob Copeland <copeland@lastpass.com>