lpass: implement 'lpass passwd'
The "lpass passwd" command (by analogy to UNIX passwd) can be used
to change the master password of a user.  Because the lastpass design
requires all encryption to be done on the client side, this involves
a download, then reencrypt, then upload of all the vault items and
ancillary data.

Signed-off-by: Bob Copeland <copeland@lastpass.com>