passwd: rework decryption error handling
Make lpass passwd work the same as the website: if decryption fails,
then clear the field completely.  If there are too many such failures
on required fields, abort the whole operation.

The previous code had a crash when calling strlen() on the null
new_ctext, but there's no point anyway in trying to not send these strings
to the backend because the server will reject the update request.

Signed-off-by: Bob Copeland <copeland@lastpass.com>