lpass: add an "add" command
While adding a new account can be accomplished with either generate
(creating the password field) or edit, some people have been confused
that there is no standalone "add" command, especially when they
are wanting to create notes or accounts that already have credentials.

So, create an 'add' which is just like 'edit' except that you can only
make new accounts with it.

Signed-off-by: Bob Copeland <copeland@lastpass.com>