blob: use account->fullname in account_assign_share
It's a bit tricky to use account_assign_share properly: because
any encryptions before the share is assigned use the blob key,
they will be incorrect after a share is assigned.  Because of
that, the API previously required that you pass in the intended
fullname of a new account before actually setting the fullname
into the account because the latter would only encrypt with the
share key after it is assigned.

Instead, teach account_assign_share to just reencrypt all the
fields if needed and pass it the blob key.  Then account_assign_share
may be called any time after the fullname is already set.

This will make it easier to fix edits that change the fullname
into a shared folder (i.e. mv without using mv command).

Signed-off-by: Bob Copeland <copeland@lastpass.com>