show: add --expand-multi flag to see all matches
Normally, when matching multiple accounts on a given search string,
lpass will print "Multiple Matches Found" and show all the matching
titles and ids, so that the user can then pick one in particular to
expand in a subsequent show command.

For some uses, it makes more sense to print the entire contents of each
site instead, so add the --expand-multi (or -x) flag to enabe this.

When multiple search arguments are given, automatically expand the
results as the condensed display doesn't make much sense in this
context.

Signed-off-by: Bob Copeland <copeland@lastpass.com>