show: fix display of ascii-armor key fields
When saving SSH keys as secure notes, LastPass will destructively
replace the newlines with spaces due to the storage format of the
note fields.  Fortunately, we can generally recognize such fields
based on their format, and undo the conversion, so do that.

This fixes an oft-reported issue when ssh privkeys are stored in
the vault; now they can be retrieved properly with something like:

    $ lpass show some-privkey --field="Private Key"

Signed-off-by: Bob Copeland <copeland@lastpass.com>