cmd: fix list corruption when performing multiple id searches
I ran into a case where find_matching_accounts was hanging.
The cause was due to a recent change to support multiple
consecutive searches on the accounts database; in order to
avoid duplicates, matches are moved from one list to another
list.  Unfortunately, in this case we forgot to delete it
from the first list so the list pointers got corrupted.

Add the missing list_del() before the list_add_tail.

Signed-off-by: Bob Copeland <copeland@lastpass.com>