share: fix crash when invalid share supplied
According to the comments, all subcmds are supposed to check args->share
before using it, but only userls actually did.  The supposed reason is
that some commands don't need share to exist, but in actuality only one
command is in that bucket: create.

At some point I had intended to move the check to common code (because share
is only looked up if !create) but did not complete that change, so do that
now.

Signed-off-by: Bob Copeland <copeland@lastpass.com>