upload-queue: tweak timeouts and give up after 5 failures
Previously we would repeat the same request five times over
a total of 45 seconds, with a max of 10 seconds between the
last two requests.  This might be too little in case a rate limit
is hit, so make this increase exponentially so that the final
attempt is over five minutes after the previous one, but
first few requests happen within a minute.

With that change, we no longer need to sleep 30 seconds after
a failed request.

If we get a server response after five successive tries,
go ahead and move the request into a separate directory
so that we won't get stuck trying the same request over and
over again.

Signed-off-by: Bob Copeland <copeland@lastpass.com>