lpass: implement support for EU datacenter
The lastpass.com server will respond with a server=X line if the
given user is actually located in the EU datacenter; in that case
the client is supposed to send all requests after that to
lastpass.eu.

It makes sense to tie that server to the session (when there is one)
because we need to remember the correct server across multiple
invocations of lpass.  Thus, we now store session_server after
successful login, and use session->server for all requests with a
valid sessionid.

Signed-off-by: Bob Copeland <copeland@lastpass.com>