export: use sharename\groupname to match website
When exporting, the website will include the sharename and
subfolder name (if present) separated by backslashes.  Fullname
isn't a good proxy for this because it uses a forward slash after
sharename, and it includes the account name as well.  So just
create the string in the right format as needed.

Signed-off-by: Bob Copeland <copeland@lastpass.com>