cipher: support new cbc format for RSA keys
LastPass is moving to an encrypted RSA key format which more closely
matches the way everything else is done: namely, instead of having
an implicit client-side iv derived from the AES key, it will have an
explicit iv that is randomly generated, and encode that in the same
format as account data.  Add support for decrypting the new format
keys.

Signed-off-by: Bob Copeland <copeland@lastpass.com>