pins: add future root / leaf keys to pin list
LastPass will soon be updating some of its certificates and
keys.  In order to maintain continuity before and after the
changes, add pins for the not-yet-in-place certificates.

As I do not yet know exactly what the chain will look like,
I've included all of the pubkeys from the probable root,
as well as the leaf certs just in case another provider is
used.  When the final chain is in place, we'll reduce this
list to a smaller set.

Backported from master.

Signed-off-by: Bob Copeland <copeland@lastpass.com>