blob: add initial support for applications
With this change, applications in the vault will show up with
lpass ls.

Applications share many of the same fields as a regular website
account, but also have a few of their own.  By embedding an
account structure in the new app container we can share these
fields and use existing account code for searching and display.

struct account gains an is_app field so it can implement app or
non-app specific code where needed.

Signed-off-by: Bob Copeland <copeland@lastpass.com>