upload_queue: use extensible arrays for params
In a lot of cases we have "if (share) add this extra thing" and so
everything else gets duplicated.  Also, when we are going to start
updating apps, the param names are rather different.  Use the argv-style
arrays instead of varargs so that we can just dynamically create this
list as we go.

Signed-off-by: Bob Copeland <copeland@lastpass.com>