show: fix crash when requested field is not found
As reported by user tod-uma on github, when a field is searched
for that doesn't exist, lpass crashes.  This is because
list_for_each_entry leaves the list cursor pointing at the list
head, not at NULL, when list iteration is complete.  Use a boolean
variable here to correct the conditional.

Fixes #167.

Signed-off-by: Bob Copeland <copeland@lastpass.com>