lpass: add CURL verbose debugging as an option
For certain failures (such as certificate errors) it is helpful
to see libcurl's take on why things aren't working.  This adds
a new debug level, verbose=8 which includes all of the libcurl
messages (including some protocol information and https payloads).

Also improve the logging infrastructure by moving it out to its
own file and using a separate stdio stream instead of a duplicated
stdout.

Most users will not want to turn on logging as the payloads can
include cookies and other sensitive data, so it will remain off
by default.

Signed-off-by: Bob Copeland <copeland@lastpass.com>