config: fix XDG defaults and update documentation
The XDG base specification specifies defaults to use in the
absence of certain environment variables.  We need to work in
environments where XDG is not used at all (e.g. on macs) but
also do the right thing if on a typical Linux box.  So check
that XDG_RUNTIME_DIR is around, and if so, then use the
defaults; otherwise fall back to ~/.lpass.

Also fix misspelled XDG_RUNTIME_DIR and update the docs
accordingly.

Signed-off-by: Bob Copeland <copeland@lastpass.com>