edit: fix Name/Name conflict when editing secure notes
"Name" in the template file could be interpreted as either the
name of the account itself, or of the (person's) name in the
template.  The result of this confusion is that editing the
Name in the SSN template didn't work; it would be assigned
to account fullname instead (which we later throw away when
editing secure notes).

Use line-number to disambiguate the case: account name should
always be on the first line, everything thereafter is part of
a note.

Also move the strdup to where it is used so the lifetime of the
string copy is clearer.

Signed-off-by: Bob Copeland <copeland@lastpass.com>