edit: do not make field entries for username/password
A user reported that using Server secure note type resulted
in some odd stuff showing up in the plugin vault.  This note
type has a "Username" and "Password" field, but we already
use that to map to account->username and account->password.

When collapsing, the contents of those fields gets copied
back into the note, along with any fields in the fields list.
Unfortunately, we *also* copied in the empty duplicate
Username/Password fields from the fields, resulting in a
note with extra fields that confused the plugin.

Fix this by not creating the duplicate fields in the fields
list in the first place; then only the version in the account
will get copied into the note.

Signed-off-by: Bob Copeland <copeland@lastpass.com>