edit: fix editing of secure note names
It was reported that editing a secure note to rename it doesn't work,
while it does work for accounts.  The reason is we only apply name
changes only if the user had specified `--name`.  But this is a
historical limitation, e.g. when using the "any" edit type it makes
sense to retain username edits, so just do it always since we
always have a copy of the original account name in any case.

Fixes #106.

Signed-off-by: Bob Copeland <copeland@lastpass.com>