ls: use format machinery for tree print as well
This drops a bit of code duplication by using the same format
string for both color-less and tree mode output.  The only
difference in the format string between the two is whether the
fullname or short acpcount name is used; we can go ahead and
embed the terminal color codes which will simply get stripped
in terminal_printf.

Signed-off-by: Bob Copeland <copeland@lastpass.com>