format: expand format specifiers, and support '/' modifier
This adds some infrastructure useful for cmd-show:

 - You can use '/' after % to include a slash only if the expansion
   of the rest of the placeholder is non-empty
 - group and shared folder name gained their own placeholders

Thus showing a path-like representation of a name can be done
like "%/S%/g%Nf".

Also, format_field function was added which formats a name/value pair
("%fn", "%fv" resp.).  It also takes an account so that it can format
any of the existing account fields as needed.

Signed-off-by: Bob Copeland <copeland@lastpass.com>