build: reinstall Makefile, to call cmake
This normalizes build to happen in build subdirectory so that
existing source tree doesn't get polluted by build artifacts.

The main targets are replicated in the top-level makefile so
that most existing build setups should mostly continue to work,
though the output files will now end up in the build/ directory.

Of course doing 'cmake . && make' will continue to put the
build artifacts in the top-level directory if someone wants
that.

Signed-off-by: Bob Copeland <copeland@lastpass.com>