test: add some tests of basic functionality
This adds a bash-driven test infrastructure that uses the lpass-test
binary to do some simple tests of login, add, edit, show, duplicate,
and generate.  All commands disable syncing and either operate on the
local blob records (of added accounts) or the accounts that are
returned in the mock getaccts function.

Signed-off-by: Bob Copeland <copeland@lastpass.com>