ls: show empty shared folders
Currently only folders with something in them get shown.
For regular folders, there is the concept of the 'group'
type account which can be used to represent an empty folder,
so these can be displayed in `lpass ls`.  For shared folders,
though, there is not such an account type.

But we know all the shared folders as they are added to a list
in the blob, so we can create a fake group account just for
displaying inside `lpass ls`.

Signed-off-by: Bob Copeland <copeland@lastpass.com>