show: display reprompt status if enabled
I plan to add editing of reprompt status, and it would be helpful
to see it in the 'show' command to see if it is currently enabled
or not.  Since the vast majority of sites are unlikely to have
reprompt enabled, only show it if set to true, to avoid adding too
much noise.

Signed-off-by: Bob Copeland <copeland@lastpass.com>