edit: add a method of editing reprompt setting
Although we track it and now show it, there is no way for a user
using the CLI to modify the reprompt setting for a site.  This
change adds a new line into the edit buffer for accounts if the
reprompt setting is set to true, like so:

    [...]
    Reprompt: Yes

If reprompt is not true (vast majority of cases), we just leave it
out.  If the value is changed to anything other than Yes during the
edit session, then reprompt is changed accordingly.  Thus reprompt
can be turned on and off like so:

    echo "Reprompt: Yes" | lpass edit site --non-interactive
    echo "Reprompt: No" | lpass edit site --non-interactive

Add a test to capture this functionality.

Signed-off-by: Bob Copeland <copeland@lastpass.com>