support new-style private key format fields
The LastPass plugins now support multiline fields (besides Notes)
if the field is part of a known note template.  This has been used
to capture the embedded newlines in ascii-armored private SSH keys
with the ssh-key note type.

Until now lpass only supported ssh keys stored on a single line,
so add the ability to parse these kinds of strings to enable
support for keys created in the newer plugins.

Signed-off-by: Bob Copeland <copeland@lastpass.com>