README: advocate top-level makefile
'make test' only works if you start with the toplevel makefile
because it now assumes cmake is run from the build/ directory
(this is the suggested practice so that build artifacts don't
contaminate the source).

So advise using 'make' instead of the cmake-generated Makefile,
at least for people who aren't familiar with cmake which will
build with cmake in the build/ directory.

Signed-off-by: Bob Copeland <copeland@lastpass.com>