blob: only do field concatenation for multiline fields
Commit 1ae6957ac2f, "support new-style private key format fields"
introduced a regression in that non-templated fields got merged
with the preceeding fields for all template types.  While this is
similar to what the website does, it broke workflows for several
users and, in practical terms, only ssh private keys have this
feature, so we only need to do it there.

Introduce a new function to tell us which template fields are
eligible for this version of expansion and only do it for those.

Signed-off-by: Bob Copeland <copeland@lastpass.com>