edit: support ssh keys in interactive lpass add
Split out the edit file parser from the processing code, and
enhance it to accumulate multiline strings in a similar manner
to expanding notes in the blob.  This allows us to put a multiline
private ssh key file in the ssh secure notes in a more natural
way than joining a bunch of lines together.

Fixes #316.

Signed-off-by: Bob Copeland <copeland@lastpass.com>