share: use a different endpoint for getting shared folder users
The old share.php getinfo endpoint doesn't return all of the
folder's user records (for example those who are in the new
unassigned key list are missing).  The 4.0 plugin uses a different,
JSON endpoint for this, so follow suit so they will match.

As this is the first endpoint to require JSON which does not also
have an xml interface, introduce YAJL to parse the response.

Signed-off-by: Bob Copeland <copeland@lastpass.com>