show: teach show --json option
Based on a patch from Pau Sanchez, add --json option for "lpass show",
which can be friendlier for certain kinds of scripting tasks.

Normal selection rules apply, so you can, for example, do:

    lpass show -jxG .*

to dump all accounts.  In the future we may switch out the custom json
code for YAJL or similar, but this gets the basics working.

Fixes #263.

Signed-off-by: Bob Copeland <copeland@lastpass.com>