tests: rename note-with-notes test to what it does
Commit be7b13012 added a test case to demonstrate a particular
parsing issue when there was a field in a secure note.

Confusingly, this field was called "Note" (unlike the actual
name, "Notes").  Rename "Note" to "Field" in order to clear up
this confusion a bit.

I verified that the original case will still fail on the original
code, whether named Field or Note.

Signed-off-by: Bob Copeland <copeland@lastpass.com>