Add Dockerfile to create a clean build environment
There are two Dockerfiles: 1) Dockerfile and 2) Dockerfile.dev

The Dockerfile is the one which has the smallest image because
everything is done in 1 RUN statement.

The Dockerfile.dev is the one that may be easier for a developer. It
caches the apt-get install phase so you don't need to rebuild everything
from scratch if you are testing your changes. And it doesn't delete the
build artifacts so you can hop into the container and see what is left
behind.

The next step would be to create a Debian/Ubuntu package based of this
Docker image. I'll leave that as a TODO for now :)

Signed-off-by: Wesley Schwengle <wesley@schwengle.net>