version.h: do not update the modification timestamp on every single run
The LASTPASS-VERSION-GEN script was copied from another project that
used an entirely different format to generate a Makefile include. As a
result, it always considered the file out of date, and always remade it,
thereby causing `make && sudo make install` to rebuild many files and
relink the executable during the install step.

Signed-off-by: Eli Schwartz <eschwartz@archlinux.org>