Decrease the time for the cli app to do things
When the local blob version and the remote version are the same skip
downloading a new blob and just continue with the current blob.
This could potentially shave 300ms (or more, I've seen close to 2
seconds in my browser) when the blob hasn't been updated on the remote
side.

There is still a small performance hit on `login_check.php` but I
cannot think of logical ways why we would increase the timeout, the 5
seconds that is place now make sense to me as you want to display
things. Maybe there should be a post action on any command that has
`--sync=auto` so subsequent calls have a blob in sync and would cause
slowness of the CLI app.

Fixes: #475
Signed-off-by: Wesley Schwengle <wesley@schwengle.net>