Fix non-default PKG_CONFIG_PATH on macOS
PKG_CONFIG_PATH shouldn't be reset, only appended to or prepended to, so
that the user can still use non-default paths if they need to.
Additionally, package managers other than Homebrew (such as Nix) still
need to be able to provide their own versions of packages, and aren't
going to use Homebrew's default path to do so.

Signed-off-by: Alyssa Ross <hi@alyssa.is>