git-credential-helper: add support for store and erase
The initial git-credential-lastpass was only answering to the 'get'
command, which meant the helper would not register new passwords,
update existing ones or remove secrets that would not work.

In this patch, I tried to lower the number of calls made to 'lpass' in
order to keep a good performance.

Signed-off-by: Maël Valais <mael.valais@gmail.com>