fix(agent.c): wait for agent to successfully spawn before exiting.
We open a pipe with the agent when we spawn it, and wait for them
to say they're ready before exiting. This *should* fix #650.

The underlying issue was that when doing:
```
lpass login foo@example.com
lpass show bar
```
in a script would cause a race condition where `lpass login` exited
before the agent was successfully brought up. `lpass show bar` then
tried to reach out to the agent, sees the socket isn't created yet,
and says the user isn't logged in.

Signed-Off-by: Arvinder Dhanoa <ArvinderDhan@gmail.com>