lpass-att-export.sh: optimize id gathering
Use an export to find the ids of item that may contain attachments. This
is better than showing every item since showing every item will prompt
for your master password for every item with a reprompt set instead of
only those we need to actually export. It will also be significantly
faster on older hardware or when the master password PBKDF2 iterations
are large.

Signed-off-by: Patrick Watson <watson81@users.noreply.github.com>