Add format string option for file-system-safety
Share and Group names are arbitrary strings and could contain characters
that are not valid on file systems such as forward slashes ('/') or
control characters. This applies to attachment file names also, though
less commonly so since they usually will have a name from the uploading
user's file system.

Thus, when exporting attachments, these strings must be sanitized. This
is painful to do in lpass-att-export.sh, so this commit introduces the
new '_' format string modifier. This modifier works similarly to the
existing '/' modifier, but causes instances of forward slashes, carriage
returns, new lines, and tabs with an underscore in output.

lpass-att-export.sh has been modified to use this new format string when
a new-enough version of lpass is present.

Signed-off-by: Patrick Watson <watson81@users.noreply.github.com>